Refactor : generate figure package
