+++ src/main/java/com/woowacourse/coordinate/domain/figure/FigureKind.java
@@ -1,4 +1,4 @@
-package com.woowacourse.coordinate.domain;
+package com.woowacourse.coordinate.domain.figure;
 
 import java.util.Arrays;
 import java.util.List;
