+++ src/main/java/com/woowacourse/coordinate/domain/figure/Line.java
@@ -1,4 +1,4 @@
-package com.woowacourse.coordinate.domain;
+package com.woowacourse.coordinate.domain.figure;
 
 import java.util.Objects;
