+++ src/main/java/com/woowacourse/coordinate/domain/figure/Rectangle.java
@@ -1,4 +1,4 @@
-package com.woowacourse.coordinate.domain;
+package com.woowacourse.coordinate.domain.figure;
 
 import java.util.ArrayList;
 import java.util.List;
