+++ src/main/java/com/woowacourse/coordinate/domain/figure/Shape.java
@@ -1,4 +1,4 @@
-package com.woowacourse.coordinate.domain;
+package com.woowacourse.coordinate.domain.figure;
 
 public interface Shape {
 	double calculateArea();
