+++ src/test/java/com/woowacourse/coordinate/domain/PointTest.java
@@ -1,5 +1,6 @@
 package com.woowacourse.coordinate.domain;
 
+import com.woowacourse.coordinate.domain.figure.Point;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
